Add Distribution.Client.Utils.Newtypes module
